+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - `ricos-editor`
   - [#1300](https://github.com/wix-incubator/rich-content/pull/1300) missing imported styles
   - [#1296](https://github.com/wix-incubator/rich-content/pull/1296) Editor Modal z-index increase
+  - [#1303](https://github.com/wix-incubator/rich-content/pull/1303) actionColor fix for "more+" button
 </details>
 <hr/>